Add audio transcript ingestion and create/upload processed files
