+++ backend/modules/ingestion/ito/summary.py
@@ -1,8 +1,5 @@
-import random
 import tempfile
-from io import BytesIO
 
-from fastapi import UploadFile
 from langchain.chains import (
     MapReduceDocumentsChain,
     ReduceDocumentsChain,
@@ -15,7 +12,6 @@ from langchain_core.prompts import PromptTemplate
 from langchain_text_splitters import CharacterTextSplitter
 from logger import get_logger
 from modules.ingestion.ito.ito import ITO
-from modules.upload.controller.upload_routes import upload_file
 
 logger = get_logger(__name__)
 
@@ -99,37 +95,6 @@ class SummaryIngestion(ITO):
 
         content = map_reduce_chain.run(split_docs)
 
-        # Now create a fake.txt file with the content of the summary with the name of the original file without the extension as an UploadFile object
-
-        content_io = BytesIO(content.encode("utf-8"))
-        content_io.seek(0)
-
-        # Create a file of type UploadFile
-
-        # Generate a new name for the file with self.uploadFile.filename without the extension and add 4 random digits at the end and then .txt
-        new_filename = (
-            self.uploadFile.filename.split(".")[0]
-            + "_summary_"
-            + str(random.randint(1000, 9999))
-            + ".txt"
+        return await self.create_and_upload_processed_file(
+            content, self.uploadFile.filename, "Summary"
         )
-
-        file_to_upload = UploadFile(
-            filename=new_filename,
-            file=content_io,
-            headers={"content-type": "text/plain"},
-        )
-
-        if self.send_file_email:
-            await self.send_output_by_email(
-                file_to_upload, new_filename, "Summary of the document"
-            )
-        # Create a file of type UploadFile
-        file_to_upload.file.seek(0)
-        await upload_file(
-            uploadFile=file_to_upload,
-            brain_id=self.brain_id,
-            current_user=self.current_user,
-            chat_id=None,
-        )
-        return {"message": "Summary generated successfully"}
